AIRFLOW/plugin/requirements.txt
pymongo
python-dotenv
pandas
numpy
python-dateutil
loguru
pytest
apache airflow